✨configurable select first file on App open
