+++ src/components/settings/index.ts
@@ -1,3 +1,3 @@
 export * from "./Settings";
 export * from "./ThemeInfo";
-export * from "./WelcomeContent";
+export * from "./ConfigToggles";
